Updated Grid
Added another grid layout to display an extended 5 day forecast of
weather details. This was mainly added to better accomodate the file
storing part of the app. Instead of caching the current condition it
will now store the 5 day cast. Incase of no connection the cached
forecast will be displayed with the text "cached" in the header to
notify the user of what type of data this is. Additionally the dates
will be shown instead of the actual day.