Finalized Provider
Added the Uri match code for the single day query. Added a scroll view
wrapper to the main view to allow all elements to be viewable. Added a
radio group for the querying types. Fixed the styling attributes of the
list view to display the data in separate columns instead of a long one.